Add cmake options
In order to support example programs, and new configure-time generated
content, new CMake options had to be added to the primary CMakeLists
file.

These options now toggle support for:
* building example projects,
* generating trig-table (for table-based math), and
* building the library as a shared library

A temporary placeholder CMakeLists file for the 'examples' project has
been added as well, to ensure that toggling this option will not fail
to configure.