Add configured header file
Using preprocessor branching dependent on macro identifiers can be
dangerous when required in headers, since a library may be compiled
with one set of defines that differs from its uses. This can have
profound effects on the ABI of the library, and can lead to various
kinds of undefined behavior, crashes, etc.

To work around this, a header file is configured during CMake's
configuration stage that serializes the set of fixed options decided
during the first build. This file is installed during the library
installation, which guarantees that the consumption of the library
will always use the same flags as the configuration/compilation of
the library -- thereby bypassing the issues listed above.