Add precision core headers
The library requires some way of toggling the default precision used
throughout the entire engine. In general, it is common for game-engines
to support:
* single-precision floating point (float),
* double-precision floating point (double),
Additionally, some systems are more limited and require either
half-precision (more embedded systems), or fixed-precision (systems
that lack FPU units, and do floating point operations in software).

This is conditionally driven from the generated configuration files.

The fixed precision and half precision headers are added as
placeholders for proper implementations, to be added when there is a
proper requirement for these types to be implemented. In the mean time,
the library strictly supports 'float' and 'double' as a single type:
'alloy::core::real' (indicating a 'real' number in mathematics).