Fix precision of real type
The 'alloy::core::real' type's precision was incorrectly fixed to
always be 'double' due to an incorrect preprocessor check for whether
'ALLOY_PRECISION_DOUBLE' was set, rather than checking what the value
of 'ALLOY_PRECISION' resolves to (as intended).

The fix is to change the definition check to instead be
'ALLOY_PRECISION == ALLOY_PRECISION_DOUBLE'.