Remove dummy.cpp file
This file was originally only added in order to make 'Alloy' a proper
library while the API was being introduced. CMake requires there to be
at least one source file in order to create a proper library, otherwise
it creates a header-only (INTERFACE) library. This source file allowed
for this to be considered a proper library, despite not actually
containing symbols.

Since the introduction of the 'trigonometry.cpp' file, this dummy
file is no longer necessary.