Rough-in precision types
Additional precision types will be made available for the alloy
library at a later date. These precision types include:
* 'half' precision type, which is a 16-bit floating point that is
  supported by modern GPUs to minimize storage capacity -- useful for
  hardware-limited systems
* 'fixed' precision type, which is a fixed-precision
  (e.g. non-floating) type that is driven by software rather than
  hardware. This is necessary for systems that do not have FPUs.

Both of these types will be able to be interchangable with the
'core::real' type in terms of all major API requirements.

The API has been roughed-in, to be defined at a later point.