Add compiler intrinsics headers
This adds new headers that abstract the underlying compiler intrinsics
behind a series of macros, in order to allow support for things that
are not part of the modern C++ language.

In order to facilitate this change, the configuration macro:
'ALLOY_COMPILER_EXCEPTIONS_ENABLED' had to be renamed to
'ALLOY_EXCEPTIONS_ENABLED', since the 'ALLOY_COMPILER' prefix is
used in all intriniscs headers.