Collapse 'math' into 'core'
With the restructuring of Alloy into several libraries and binaries,
the namespacing and subsystem for the 'alloy-core' library has been
collapsed into the single 'core' namespace and subsystem.

This restructuring is motivated by the desire to keep project names
and subsystems in-line with namespaces. This leads to issues if
the public include becomes 'alloy-core', since namespaces cannot
contain the '-' character. At which point, it's ambiguous whether
it should be 'alloy::core', 'alloy_core', etc. This leads to more
difficulty when the 'core' subsystem is considered, since it would
become 'alloy::core::core', or 'alloy_core::core', or even just
folded directly into the root 'alloy::core'. This also allows for
the include to remain 'alloy/core'.

Various alternatives were explored, including the ones mentioned above,
but ultimately the best presented solution was to merge everything
into the 'core' library directly, and to make the library's purpose
cohesive. The purpose of the 'core' library is now to define the
value and semantic types that are required for a game engine, including
all mathematical types, graphical primitive types, etc.