Move math types to 'core/math'
This moves the math types, previously in the 'alloy/math' subsystem,
into 'alloy/core/math' for better containment.

This decision came after much deliberation as to the proper structure
for the alloy-core library. The decision was taken for only the
top-level directory to dictate the namespace, thus 'alloy/core'
defines 'alloy::core' -- but 'alloy/core/math' does not become
'alloy::core::math'.