Fix broken is_vector utility
A previous change which removed the templates from the various
vector types unintentionally broke the 'is_vector' by not removing
the 'typename T' arguments from the template. This results in
compile-errors, since the 'T' type is not deducible.

The fix is to change 'template<typename T>' to 'template<>' for the
various 'is_vector' specializations.