Add 'ray' geometric type
This introduces the 'ray' geometric type, used for line-of-sight
calculations in 3d graphics.

The 'ray' type is purely a semantic type wrapper around a point
and a direction -- but does not contain any logic for detection.
This will be coming in a later update that will manage all intersection
of geometric primitives.