Add clang-format and run script
This commit introduces the first version of a clang-format style
to this repository.

NOTE: This format guide is not yet complete, and may require several
alterations. Because of this, the run-clang-format.py script that
has also been added to this commit does not destructively overwrite
files -- it only prints to console so that it may be used as a
guideline (for the time being).

Once the formatting details has been properly tuned to this
repository's needs, the run-clang-format script will be updated to
reflect the new changes.