Clean up gitignore
The gitignore has been significantly cleaned up.
It now only lists files known and expected to be generated by this
repository, rather than creating a comprehensive list of possibly
generated files based on the tools that a user might (or might not)
be using.

It is now the responsibility of the author to have these types of
generated content blocked by a personal setting.