Clean up tools scripts
The 'run-doxygen' and 'run-clang-format' scripts contain a function
'root_path', which has a variable in the function's scope also named
'root_path'. This causes the variable name to shadow the function name
when within the function scope -- which additionally hampers
readability.

The fix is simply to rename 'root_path' to 'path'