Fix source-glob location in run-clang-format
The run-clang-format script was only globbing for sources from the
old 'include' directory -- and not even looking in any 'src' directory.

Since the paths have now been restructured to be under various
subprojects under 'lib', 'extra', 'example', and 'bin' -- this glob
needed to be updated to search under any of these locations.