Conditionally compile Alloy::IO::SDL2
Alloy::IO::SDL2 is now conditionally compiled based on a
cmake-dependent option that uses ALLOY_COMPILE_EXTRAS and the existence
of SDL2 as option requirements.

This will be compiled by default if ALLOY_COMPILE_EXTRAS is enabled and
the SDL2 library is found, otherwise this option will not present
itself. This will be a general pattern moving forward.