Update sdl windows to follow best practices
The sdl window type constructors have been modified to only accept the
complete set of members as the primary constructor. This is considered
best practices in many languages (rust,swift,etc) since it provides
complete initialization.

The previous working-constructor logic has been changed to named
static-functions instead, to provide a means to explicitly state
construction logic.

This practice is already followed through the majority of the rest of
Alloy, and will be adopted as *the* proper development practice for
the future.