Remove version.hpp.in
This header defines the Alloy project version, and was previously
generated into the alloy-core library's headers. However, the version
is static across *all* of alloy and not just the core, and as such
should be a product of the entire library and not just parts.

This will be added back in the future, likely to the 'Alloy::Alloy'
target which will include all library components.