Rename 'NOT_SUBPROJECT' variable in CMakeLists.txt
This renames the variable which is used to indicate whether this
cmake project is included as a sub-directory, or built from the
top-level. The previous name 'NOT_SUBPROJECT' necessitate the use of
double-negations to check for whether this is a subproject, which
lead to not-so-nice code: 'NOT NOT_SUBPROJECT'.

As a result, this has been renamed to 'ALLOY_IS_SUBPROJECT'