Fix broken use-discrete-gpu project
The use-discrete-gpu project was broken by the removal of the
API macro definitions from 'config.hpp'. Since these macros are
defined on a per-library basis now, there is no simple header to
include these macros.

However this project produces a static library that never exports
the symbol, and so the change was to outright define the macros
as part of the source, rather than include from another source.