Add missing header to vector_utilities.hpp
vector_utilities was incorrectly checked-in in a broken state. This
breakage was introduced with the recent changes to the vector classes.

As part of the refactorings, headers were changed to no longer include
the vector classes as it previously did. Removing these headers
exposed a preexisting problem that this header never included
the definitions for the `alloy::core::real` type -- which now causes
this file to break during header self-containment tests.