Update `sdl2-bindings` to use new semantic types
The `sdl2-bindings` extra project has been updated to only use
the new `window::dimensions` and `window::position` semantic types,
rather than supporting between 2 to 4 `ints` in sequence.

This helps add clarity to the callsite of consumers of the API, and
helps to prevents accidentally passing (x,y) positions to the
(width,height) receivers and vice-versa.