Post `window_close_event` calls as immediate
The `window_close_event` calls were previously posted as normal
events with regular precedence -- which could potentially not be
acted on immediately by a listener to the `message_pump` if it
were to be handling events either asynchronously, or handling them
via priority.

To potentially circumvent this, the `window_close_event` is now
handled via `post_immediate` to provide a higher precedence.