Make immediate events handled by default handler
The `message_pump::listener` type previously had default
implementations for both the `handle_message` and
`handle_immediate_message` functions that would discard the event,
doing nothing.

This poses a bit of an anti-pattern, since a listener by default will
have the default behavior of doing nothing until overridden; and a
listener may never receive immediate messages.

To work around this, the listener has been updated so that the
`handle_immediate_message` handler will delegate to the
`handle_message` handler, and the `handle_message` function no longer
has a default implementation -- which forces the consumer of this API
to provide a "do-something" implementation.