Add component_manager type
This adds a new manager for any arbitrary T component type. The
new manager uses templates and simple type-erasure to allow for
arbitrarily adding new 'component' types to be managed that can be
associated to any entity type.

NOTE: this type is not fully integrated yet. It succeeds in compiling,
but may not be in a fully functional state as it has not been
completely tested yet.