Add 'expected' utility type
This type introduces a new form of error handling that is orthogonal to
the standard exception handling. Rather than throwing exceptions, or
even just returning an error code, this simplifies the effort by
returning a dual-state object that may be either in an error condition
or a value condition.

This will be used around the system to provide 'noexcept' guarantees
rather than requiring exceptions