Update component_manager to use type_index
The component_manager previously had an implementation of type_index
done inline in the class for providing unique indices of component
types. This approach needed to be more broad to be accessible to other
classes as well, which is why type_index was created.

As a result, the previous component_info has been removed and replaced
with the new type_index, since it directly supercedes this.