Add lightweight delegate type
This new type exists for rebinding raw function pointers, and pointers
to member/member-functions in a simple, lightweight class.

Due to the way this class operates, compilers like gcc and clang
(presumably others like MSVC as well) are able to see into the
types and optimize many of the indirect calls -- making this superior
to std::function in most regards. The lack of heap allocations also
simplifies things greatly.

The only complication that occurs from this is that delegate requires
a static template function to be invoked in order to produce a delegate
type -- rather than binding easily in the constructor. This effectively
takes 2 steps rather than std::function's 1 step in order to bind
functionality -- which ultimately is a small price to pay.