Add observer_ptr
This new utility allows for having a simple, non-heap way of observing
a given type with weak viewing semantics. This effectively implements
a weak_ptr that does not require a heap-based node, and rather uses
an intrinsicly linked list to store all observables to an object.

This creates a really simple way to view types that might otherwise
not be on the heap. This becomes powerful and useful for single-threaded
observers that might expire (potentially being forgotten to be
unregistered).