Add more restrictions to allocator's make/dispose
More static_asserts have been added to allocator::make, make_array,
dispose, and dispose_array to prevent misuse.

Only value-like types, not references or array types, may be
created (and thus, also disposed) from allocators -- as it would be
nonsensical and an error to allocate references (which are required
to be bound to a referee).