Add mutable and const byte buffers
Buffers have been added for both mutable and immutable sequences of
bytes. Buffers are lightweight, copyable, non-owning objects that refer
to a segment of memory -- usable in communication APIs (files, networks,
etc).

This will be required for integration into the filesystem API, as well
as potentially useful for a future networks API.