Add file_stream_handle
This handle is now used to wrap the stream and the finalizer together
into a single RAII type, to allow adapted file types to not have to
directly interact with the `file` API.

Previously, adapted `file` types were composed of other files; but this
resulted in invariants being checked multiple times (e.g. file::read
resulting in checking if the file was closed N times for N adapted
files).

With the new stream handle, only the outer `file` instance checks the
invariants, and the handle can be free to assume that these invariants
are upheld.