Fix xor_file_adapter's "write" implementation
The "write" implementation of xor_file_adapter was unintentionally
discarding the result of one of the XOR passes, which resulted writing
being asymmetric from reading. This caused the reads to convert the
raw bytes into a different stream than was written, making the adapter
effectively useless.

The unintentional discard has been fixed, which now fixes the behavior.

Additionally, the read/write calls have been given better documentation,
and now also computes the total amount of bytes written, which may
differ from the total requested bytes. This was a minor bug possible if
the write size was greater than the internal buffer size (128 bytes),
where the last chunk being written would fail (but would return a
buffer claiming it was successful). Not a critical bug, but one that
has been fixed now nonetheless.

The buffer size used in this adapter might need to be tweaked for
better performance. 128 bytes writes is pretty small.