Flush console logger after each write
std::printf should be buffered on the existance of newlines, but
curiously this does not work correctly when using MINGW -- causing log
messages to be delayed until such point that std::printf decides to
flush.

To provide consistency on different systems, stdout is flushed after
every print statement to ensure that new messages will always
visible on the console. This is necessary in case the system crashes,
to help trace the origin of the error.