Detach remaining log streams on logger destruction
It is possible for the logger to outlive log streams that are attached
if the log streams were constructed before the logger. This could
result in undefined behavior, though in general is more of a usability
error/issue than anything else.

This issue has been addressed from within the logger itself by
detaching all log streams before destructing the logger, so that when
the log stream destructs it will not attempt to detach itself as well.