Fix stl_allocator's Allocator-concept compliance
The stl_allocator did not correctly mark the copy/move/swap propagation
requirement on the interface, which is breaking default move assignment
operators downstream due to the requirement of deallocating memory
prior to allocating a new chunk of memory to move code into.

Additionally, libc++ seems to improperly require the 'rebind' template
in the allocator (rather than using allocator_traits::rebind_alloc),
and this breaks some uses of allocator rebinding.

By adding these traits, this should fix broken move-assignment operators
which are failing to build due to not being deduced as noexcept.