Add Alloy::Extra::Allocators project
This is a new project that will house all custom memory allocators for
the Alloy library. This is an optional component, since memory
management is not necessary for all alloy projects.

The project is currently given a 'dummy.cpp' file to ensure that the
library built is static, even though it currently only contains a
header for pointer utilities. This will be removed as real sources
begin to get added to the project.