Fix 'from_object' factory for mutable/const buffer
Both the io::mutable_buffer and io::const_buffer suffer from the same
bug (due to copy/paste) wherein the from_object factories incorrectly
compute the size of the object. This is due to the delegation to
'from_data' which takes the number of entries in the 'data' pointer
rather than the size of the data pointer.

The bug comes from the from_object implementation passing sizeof(T)
to the count parameter, when it should simply be '1u' since
'from_data' does the byte computation of 'sizeof(T) * count'.