✨ Add function_traits type traits
There are several cases where it is useful to extract the nth parameter
type from a function, or the return type from a function. To make this
possible, a `function_traits` type along with a `function_type_traits`
type has been added. These traits provide a simple interface for
extracting the arity of the function, the types of each parameter,
and the return types of all pointer types (function pointers,
pointer-to-member-functions, etc).

In order to make this possible, a separate related trait, `select_nth`,
was added to allow extracting the Nth parameter from a variadic set.