✨ Add source_location
Several utilities require a means of identifying the current source
location. This can be done through a simple macro that expands the
__LINE__/__FILE__ and uses the __func__ builtin; but this requires the
context to always be created in a new macro.

In order to simplify this, the source_location (which is due to be
standardized in C++20) was added, with a macro
`ALLOY_CURRENT_SOURCE_LOCATION` to always extract out the current
source location information.

This now allows this to be used in other compound utilities, such as
logging, assertions, etc.