✨ Add std::byte* traversal functions
Since the 'bytes' semantic type is intended to measure quantities
for byte buffers, it would be useful to be able to allow traverals to
take place across contiguous bytes so that expressions of the form:

```
p + kibibytes{1}
```

would be a well-formed expression to produce a pointer `p` that is
offset by 1024 bytes.

This has been added by simply adding operators that act on 'bytes' so
that any units larger than 'bytes' will be automatically converted to
the appropriate size prior to doing the traversal.