✏ Rename 'capacity' to 'data_quantity'
The original name for the 'bytes' semantic type, 'capacity' was a bit
of a misnomer. Though a capacity of a buffer is in units of bytes, it is
more correct to refer to this as a quantity of data
(or 'data_quantity').

This change allows for the translation operators for std::byte*
pointers to have a proper semantic meaning; since scaling a capacity
passed a point makes less sense than scaling a quantity passed a point.