🗑️ Remove 'octets' type
Unfortunately, the semantics of the 'octets' quantity type was not
fully thought through.

The intention was originally to allow for bytes of sizes other than 8
bits to be convertible to 8-bit octets. However, the fundamental
design of the 'data_quantity' class has a ratio of '1' being the base
'byte' unit. This is also representable as 'CHAR_BIT/CHAR_BIT' (which
reduces to 1/1). Similarly, this also means that an octet would be
represented as '8/8' (or '1/1'); which means that the representation
would never be convertible without a rethinking of design to
something else.

To avoid any of the complications that otherwise come from awkward
edge-cases for a scenario that is unlikely to be hit in practice,
the 'octets' series of quantities has simply been removed.