✨ Add string_view and zstring_view
The C++17 standard type std::string_view is great for representing
non-owning strings, such as static character sequences ('c-strings').

The downside to this type is that it lacks any way of conveying whether
a string is a null-terminated string (a true 'c'-style string),
rather than a sequence of possibly not null-terminated characters.

It's important that this distinction be made when dealing with legacy
C APIs that exclusively deal with null-terminated C strings.

To account for that, a 'zstring_view' type has been added to the
Alloy library that is only constructible from null-terminated strings.
This will allow safe interoperability with C APIs that require null
termination.