📇 Add commit standards documentation
Since Alloy is now following a new emoji-key-guided commit message
scheme, it would be helpful to share this scheme so that potential
contributors may follow it as well, and so that consumers of this
repository can have an easier time following the message scheme.

To allow for this a `commit-standards.md` file has been added to `/doc`,
along with a link to it from the `CONTRIBUTING.md` file to allow for
better visibility.