✨ Add hashed_string_view
In game engines, it can often be useful to have resources that are
identified by integral values, but correspond to physical resource
paths (e.g. files).

To accomplish this, a new hashed_string_view type has been added that
can handle string literals and provide a (possibly) compile-time
generated hash of the string path so that it may be used as an
integral value.

To enable this functionality, a new hasher -- the fnv1a_hash -- was
added, so that any arbitrary strings may be hashed.