🔨 Make entity id type configurable
The previously set entity id type of 'std::uint16_t' is lacking in that
it only supports, at most, 2^16 entities in total -- which may not be
a large amount overall when considering the number of entities in a
normal game.

To enable this, a new configuration `ALLOY_ENTITY_TYPE` has been
added to the `CMakeLists.txt`, which generates a new configuration
header, `alloy/engine/config.hpp`. This header now defines the
`ALLOY_ENGINE_ENTITY_TYPE` that is used internally.