🔨 Update delegate binding API
The delegate API defies the current standard for naming conventions
with its static factories. Rather than having the factories being named
`make` or `from_...` to provide a fluent API, the factories were named
`bind`. This is not an error, but it is inconsistent with other
APIs and sounds a little misleading, since `bind` sounds more like a
mutator rather than a static factory.

The factories have since been renamed to `make` for consistency, and
new member functions have been added for `bind` which modifies existing
delegates in place.

As an added extension, a new function `reset` has been added which
unbinds the bound function as well, setting it back to a null pointer.