🔨 Overhaul signal class
The signal class has been significantly overhauled. Rather than
inverting control and requiring the consumer to allocate space for
'sink' and designate all handlers, the 'sink' now acts as the
consumer-facing registrar for the signal.

From this also now comes a 'connection' and 'scope_connection'
class which provide a means to disconnect the connection at any point.