🩹 Fix obsolete use of delegate
The optional/extra logger implementation was making use of the
`delegate::bind` static member function -- which was changed recently
to be `delegate::make` with `bind` being a member function.

This causes build failures only when the logger project is enabled,
which is how this was missed. The use of `bind` has been updated to
be `make`.