🩹 Fix bad assertion condition
An assertion condition in the `signal` class mistakenly had the
condition inverted. The correct condition should be that a returned
iterator from `std::find` should be the end iterator (since we don't
want a listener registered); the incorrect one that was here was
checking that it wasn't the end (which would imply it exists in the
list). This error would cause connection of listeners to always fail.

The code has been corrected to reflect the proper invariant.