🗑️ Remove unused axis_aligned_box function
axis_aligned_box declared, but did not define, the function 'area'
which was to return the total area of the axis-aligned box.

It was a bug to not define a declared function, however there is not
a hugebenefit to defining this function at this time.
To keep the API surface area simple, this has been removed.