🔧 Add enable_testing to tests
In order to properly use 'ctest' from the root build directory,
the 'enable_testing' command must be invoked from the highest-level
CMakeLists.txt. Without this call, ctest is unable to discover tests
from a directory tree.

This call was previously guarded by a condition that never enabled it.
This condition has been fixed so that it works correctly.