🩹 Remove cachelines from data_quantity
The cachelines object from data_quantity resulted in several problems
in practice. In general, it seemed to not be implemented in several
compilers, leading to this type to not be implemented either.

It's likely more convenient for users to define this themselves if this
is needed, and as such has been removed.