🔨 Convert '__LINE__' to 'size_t'
The ALLOY_ASSERT macro internally uses unsigned values for line
numbers, yet __LINE__ is signed. This leads to some compilers
complaining about implicit conversions when being extremely
pedantic.

In general, it is safe to assume that __LINE__ will never be negative,
and as a result is simply static_cast-ed to size_t