✨ Add aligned_storage
Visual Studios makes writing cross-platform code extremely difficult,
since they willfully choose to not implement the C++ standard
correctly, and keep it this way for ABI compatibility.

The std::aligned_storage type was not implemented correctly for
over-aligned storage. To get proper standards-conforming behavior,
one has to define the macro _ENABLE_EXTENDED_ALIGNED_STORAGE before
including <type_traits>

The problem is, for this to be used in a header, it requires that all
possible inclusion scenarios define this FIRST before including the
header -- including if used in a transitive inclusion graph. This is
impossible for a public type in a library header that consumers may use,
since the only way to guarantee stability is to force the consumers to
define the macro with a /D argument rather than as a #define.

Rather than fight with this hassle, std::aligned_storage_t has been
implemented inline here as alloy::core::aligned_storage.