🩹 Fix bad pragma
A clang diagnostic pragma was mistakenly defined using
'diagnostics' instead of 'diagnostic' (note: singular in correct
version). This resulted in failures with clang on windows, but not
with clang on linux/mac. This is due to the #if defined(__GNUC__)
check occurring first -- which matches clang on those builds.

The order of branches has been restored to perform __clang__ first
before checking __GNUC__, and the pragma 'diagnostics' has been
corrected to 'diagnostic' so that this will function correctly now.