🔨 Define keyword for MSVC
MSVC once again likes to play with non-standard implementations of the
C++ standard. In this case, they chose to _deprecate_ std::fopen --
which results in warnings (and thus, failures with -Werror) on Windows.

To work around this, the symbol _CRT_SECURE_NO_WARNINGS must be
defined prior to including cstdio.