📖 Prevent used-but-marked-unused warning
The ALLOY_ASSERT macro appears to trigger a warning on some versions of
clang for its use of fprintf, which warns that fprintf is marked unused
but is being used. With -Werror, this triggers compile-time errors.

Given that the assert is part of a header that consumers may use, this
is seriously detrimental. This has been temporarily suppressed while
the issues is looked into.