📖 Suppress floating point warnings
Several parts of the math library warn due to floating-point
precision -- either on loss-of-precision or due to direct comparisons.

Most of these warnings are primarily informative -- letting the
consumer know that exact comparison of floats is often a source of
bugs. However, given that the library intends to provide both exact
_and_ approximate solutions -- this is intentional, and thus suppressed.
If a consumer uses these operators with warnings high enough to trigger
these errors, it will still trigger; but it will not cause failures when
compiling the library or using the library so long as these are not
used.