📖 Suppress 'default' case in switch warning
Some compilers warn when 'default' is used as a case for a
strong enum type, despite the enum not logically being able to reach
this value. However, in C++, it is always possible for enums to
hold state other than the discriminated values.

This has been suppressed, since it's in a header, in case any other
consumers have warning levels capable of triggering this.