🔧 Disable unuseful warnings for clang-cl
clang-cl's -Wall maps, strangely, to clang's -Weverything -- which
triggers some hyper-verbose errors. Some of these errors are, in the
long-term, quite reasonable to cover -- since they are in headers, and
if a consumer uses this library they may also trigger the warnings
transitively (or errors with -Werror).

However, many of these are just not useful, or even more detrimental
than they are worth -- since they would require all builds to markup
completely valid code to prevent the warnings from occurring.

For example, exit-time destructors are a necessary evil to implement
the error_category system from the standard library; yet this triggers
warnings. Either all error categories require the same repeat pragmas
to prevent this; or we simply disable it here.