🔨 Ensure macros are not empty statements
The ALLOY_ASSERT and ALLOY_ASSUME macros both expand into either an
empty statement or a proper expression.
When it's an empty statement, it triggers empty-statement warnings with
some compiler warnings due to the semicolon.

This has been replaced with a (void) 0 to ensure that the statement
is well-formed but does not perform any action.