🔨 Pin build directory for Windows builds
When building using Ninja on Windows, for some reason ctest is unable
to detect unit tests -- resulting in no tests running. This occurs both
locally and on Github Actions CI workflow -- which is particularly
bad, since this means that tests aren't be executed when they should be.

Investigations seemed to indicates that Windows' Path environment
variable was not being updated with the build directories of the
dependent *.dll files (despite an appropriate target_link_library
being set). This causes the unit tests to not link correctly when
interrocating the executable for all tests -- which in-turn causes ctest
to think no tests are available. All manual efforts of updating the
paths seem to have failed.

The simplest fix for this is to simply pin the build directory so that
all binaries are stored in the same path alongside one another.
This is ultimately the approach that has been chosen to address this
situation. When building unit tests on Windows with BUILD_SHARED_LIBS
set to true, the path will be pinned. This does not seem to be
needed in any other executable environment, and is thus guarded.

Fixes #7