✨ Add coverage build type
Code coverage builds are an orthogonal build type to debug and release
builds, since many of the flags are similar but not identical to
the typical Debug build. This divergence necessitates a full rebuild
of the project with the specialized flags -- and so this has been
introduced as a new orthogonal build type.