✨ Update assert visibility
Compiler intrinsic breakpoints do not correctly trigger breakpoints in
the debugger if it gets triggered from within a loaded shared library.

The previous definition of `ALLOY_ASSERT` defined the assertion
function as beng `inline`, which allows the compiler to pull this
symbol from any loaded shared library. This could allow `ALLOY_ASSERT`
to be loaded from `alloy-core.so`, and may result in the breakpoint
not correctly being triggered.

In order to avoid this case, the breakpoint has been given explicit
`internal_linkage` with private visibility, to ensure that the
symbol will be inlined as much as possible.