✨ Add clamped semantic type
It's often desired in graphics programming to be able to represent
values between [0, 1]. This could be done by simply using
raw `float` values, but this doesn't convey itself well on APIs.

To represent this, a new `clamped` type has been added. This type
simply checks that the value is normalized between [0, 1], using
assertions to guard creation outside of this range.