✨ Add saturated type
The previous commit introduced a `clamped` type for forcing floating
point values between the range of [0, 1] without performing any
saturation logic.

This was previously done to ensure lightweight floating-point
operations. However, in practice, there are often cases where saturation
logic is desirable so that arithmetic operations will saturate to either
0 or 1. This logic is not desirable for all cases since it imposes
branches on all operations, and is thus not the default to the
`clamped` type.

Instead, this has been added to a new type -- a `saturated` class.
This acts like a runtime saturated floating point value -- effectively
the same as a `clamped` value that constantly clamps after every
operation. This allows for the weight of saturation logic to be
opted-in when desired, while `clamped` objects may be used otherwise.

The `clamped` class has also been given a custom constructor
that allows it to be constructed directly from other `saturated` types,
since the input range is identical (between 0 and 1).