✨ Add intersection utility
Geometric intersection calculations can frequently yield several
different geometric results for a single calculation. For example, an
intersection between two lines may be result in nothing, a point, or
a line.

In order to represent this form of intersection, a new semantic type
has been added for `intersection`, which will be used in intersection
utilities.