✨ Micro-optimize unnecessary branch in delegate
Part of the desired API behavior of the `alloy::core::delegate` is for
an unbound or reset delegate to `throw` an exception, or to at least
assert and terminate the application.

After doing some benchmarking on large volumes of bound delegates, it
appears that this frequent branch must be tested on all invocations of
the delegate, and that the compiler does not optimize this out with
`-O3`. This is not, strictly speaking, bad; however in an event-based
system, this may affect branch prediction long-term if these branches
are inconsistent, which will reduce throughput.

Thankfully there is a simple fix to it: Make the default "null" state
of the delegate actually be bound to a non-returning function that
will unconditionally throw or assert. This ensures that all bound
delegates will be able to operate with the same level of throughput that
a normal function pointer will, without the extra cost of branch
detection.

In the event that a function is unbound, the delegate will call the
default handler -- which will either throw an exception or terminate
the application accordingly.