✨ Add compiler::assume_not_null
There are many cases where the compiler will generate suboptimal code
because it must first assume that a pointer may be null. This may occur
when a function is passed to arbitrary templates that may itself perform
a nullability test first, or may sometimes just exist to attempt to
determine fallability before a dereference in some compilers.

To attempt to prevent these branches from being instantiable, a new
intrinsic has been added -- `compiler::assume_not_null`. This behaves
much like the `assume_aligned` in that the returned pointer will be
marked in such a way that the compiler is free to generate more optimal
code. This can ensure that instantiated branches will not be taken since
the compiler is already aware that the underlying pointer cannot be
null.