🔨 Refactor delegate to work with clang-cl
clang-cl for some reason is unable to support lambda expressions whose
arguments are based on variadic template arguments from the surrounding
function. Why this is the case is not known... But this poses a
portability problem to working with clang-cl any time a `delegate` is
used, since this pattern is used frequently inside the class.

To address this, the delegate class now uses static functions for each
of the stubs, rather than the slick lambda trick -- so that the lowest
common denomnator compiler will still continue to function.