✨ Add quantity semantic type
Quantities of objects in C++ development is traditionally done using a
simply `std::size_t` object, or a `size_type` alias representing some
`unsigned` integral value.

There are several pitfalls to this:

* The types do not convey, from the type-system, what they are a
  quantity of
* The types are `unsigned`, which can lead to unintentional arithmetic
  errors
* "size" objects of different unrelated values can be intermixed easily,
  which can cause unintentional runtime errors

From a topological perspective, it would be much cleaner if the
concept of a "size_type" actually conveyed one simple thing: The type
of object for which its a quantity of.

This is what the `quantity` object hopes to reduce -- since it encodes
the type of the object being counted into the quantity, and restricts
operations to prevent unintentional operations and general user
errors.